// src/enums.ts
export enum StatusEnum  {
    ACTIVE = 'Active',
    INACTIVE = 'Inactive'
};

export enum DepartmentsEnum  {
    PE = "Product Engg",
    QA = "Quality Analyst",
    DT = "Data Analysis",
    UIUX = "UI/UX",
    IT = "IT",
    HR = "HR",
};

export enum RolesEnum  {
    DEVELOPER= 'Developer',
    QA_ENGINEER= 'QA Engineer',
    CUSTOMER_SERVICE_MANAGER= 'Customer Service Manager',
    PROJECT_MANAGER= 'Project Manager',
    UX_UI_DESIGNER= 'UX/UI Designer',
    DEVOPS_ENGINEER= 'DevOps Engineer',
    SYSTEM_ARCHITECT= 'System Architect',
    BUSINESS_ANALYST= 'Business Analyst',
    HUMAN_RESOURCE_MANAGER= 'Human Resource Manager',
};

export enum LocationsEnum {
    HYDERABAD = 'Hyderabad',
    BANGALORE = 'Bangalore',
    CHENNAI = 'Chennai',
    MUMBAI = 'Mumbai'
}